# anime-api-client/src/main/resources/openapi.yaml
openapi: 3.1.0

info:
  title: API de animes
  version: 1-oas3.1

servers:
  - url: http://localhost:8080
    description: Generated server url

paths:
  /producers:
    get:
      tags:
        - producer-controller
      operationId: findAllProducers
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProducerGetResponse'

    put:
      tags:
        - producer-controller
      operationId: updateProducer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProducerPutRequest'
        required: true
      responses:
        '200':
          description: OK

    post:
      tags:
        - producer-controller
      operationId: saveProducer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProducerPostRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProducerPostResponse'

  /animes:
    get:
      tags:
        - anime-controller
      operationId: findAllAnimes
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnimeGetResponse'

    put:
      tags:
        - anime-controller
      operationId: updateAnime
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnimePutRequest'
        required: true
      responses:
        '200':
          description: OK

    post:
      tags:
        - anime-controller
      operationId: saveAnime
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnimePostRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AnimePostResponse'

  /producers/{id}:
    get:
      tags:
        - producer-controller
      operationId: findProducerById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProducerGetResponse'

    delete:
      tags:
        - producer-controller
      operationId: deleteProducer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK

  /animes/{id}:
    get:
      tags:
        - anime-controller
      operationId: findAnimeById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AnimeGetResponse'

    delete:
      tags:
        - anime-controller
      operationId: deleteAnime
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK

  /animes/paginated:
    get:
      tags:
        - anime-controller
      operationId: findAllAnimesPaginated

      x-spring-paginated: true

      parameters:
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0

        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1

        - name: sort
          in: query
          description: >-
            Sorting criteria in the format: property,(asc|desc). Default sort
            order is ascending. Multiple sort criteria are supported.
          required: false
          schema:
            type: array
            items:
              type: string

      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageAnimeGetResponse'

components:
  schemas:

    ProducerPutRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
          x-field-extra-annotation: "@jakarta.validation.constraints.NotNull(message=\"The field 'id' is required\")"

        name:
          type: string
          minLength: 1
          x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank(message=\"The field 'name' is required\")"

      required:
        - id
        - name

    AnimePutRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
          x-field-extra-annotation: "@jakarta.validation.constraints.NotNull(message=\"The field 'id' is required\")"

        name:
          type: string
          minLength: 1
          x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank(message=\"The field 'name' is required\")"

      required:
        - id
        - name

    ProducerPostRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank(message=\"The field 'name' is required\")"

      required:
        - name

    ProducerPostResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64

        name:
          type: string

    AnimePostRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank(message=\"The field 'name' is required\")"

      required:
        - name

    AnimePostResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64

        name:
          type: string

    ProducerGetResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64

        name:
          type: string

        createdAt:
          type: string
          format: date-time

    AnimeGetResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64

        name:
          type: string

    PageAnimeGetResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64

        totalPages:
          type: integer
          format: int32

        size:
          type: integer
          format: int32

        content:
          type: array
          items:
            $ref: '#/components/schemas/AnimeGetResponse'

        number:
          type: integer
          format: int32

        first:
          type: boolean

        last:
          type: boolean

        numberOfElements:
          type: integer
          format: int32

        pageable:
          $ref: '#/components/schemas/PageableObject'

        sort:
          $ref: '#/components/schemas/SortObject'

        empty:
          type: boolean

    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64

        unpaged:
          type: boolean

        pageNumber:
          type: integer
          format: int32

        pageSize:
          type: integer
          format: int32

        paged:
          type: boolean

        sort:
          $ref: '#/components/schemas/SortObject'

    SortObject:
      type: object
      properties:
        empty:
          type: boolean

        unsorted:
          type: boolean

        sorted:
          type: boolean

  securitySchemes:
    basicAuth:
      type: http
      scheme: basic